Adds the code in c#
transforms the code I add into C#, will see if I need to create it from 0 because chatgpt may be overcomplication the logic, I think the main problem is that its made so that it is always a single order which is not idea, the scheduler should always be the same, the difference would be how it is printed and how the different instructions are sent, what happens if two instructions of the same cycle start in the same one? will I have to print the differend retired? will check the powerpoints